subdivide tournament page into client component
